Fix (some) Java 21 warnings
Hopefully the uncontroversial ones .. the rest will be fixed in another
PR.

+++ addOns/tokengen/src/main/java/com/fasteasytrade/JRandTest/Tests/BinaryRankTestFor6x8Matrices.java
@@ -175,7 +175,7 @@ public class BinaryRankTestFor6x8Matrices extends Base {
 					row[j] = uni();
 				else
 					row[j] = (uni() >>> rt);
-				row[j] &= mask;
+				row[j] &= (int) mask;
 			}
 
 			cls = rankBinaryMatrix();
